Extracting kore2K parsing functionality
